Bulk Copy Return Rows (#2150)
Added int return value to NpgsqlBinaryImporter.Complete().  This will
show the number of rows that a bulk COPY inserted.

Fixes #2112 .

+++ src/Npgsql/NpgsqlStatement.cs
@@ -35,7 +35,7 @@ namespace Npgsql
         /// See the command tag in the CommandComplete message,
         /// http://www.postgresql.org/docs/current/static/protocol-message-formats.html
         /// </remarks>
-        public uint Rows { get; internal set; }
+        public ulong Rows { get; internal set; }
 
         /// <summary>
         /// For an INSERT, the object ID of the inserted row if <see cref="Rows"/> is 1 and
